fix build for beakers

+++ test/test-time_now_str.cc
@@ -1,3 +1,5 @@
+#include <string.h>
+
 #include "base/all.h"
 
 using namespace base;
@@ -11,7 +13,7 @@ TEST(misc, time_now_str) {
         time_now_str(now);
     }
     t.stop();
-    EXPECT_EQ(strlen(now), TIME_NOW_STR_SIZE - 1);
+    EXPECT_EQ((int) strlen(now), TIME_NOW_STR_SIZE - 1);
     Log::info("time_now_str() %d times takes %lf sec, that's %.2lf usec per op",
         n, t.elapsed(), t.elapsed() * 1e6 / n);
 }
